+++ escoba/src/store.js
@@ -75,7 +75,7 @@ export default new Vuex.Store({
     }
   },
   actions: {
-    loadDeck: function ({ commit, state }) {
+    loadGameData: function ({ commit, state }) {
       axios.get('http://127.0.0.1:5000/makedeck')
         .then(function (response) {
           console.log(response)
